Update aiohttp version and refactor api modules. Cleaned makefile.
