+++ docker/api/Dockerfile
@@ -3,7 +3,8 @@ ARG PRJ_PATH
 WORKDIR /usr/noetl
 
 COPY  ${PRJ_PATH}/noetl/api.py \
-      ${PRJ_PATH}/noetl/gql.py ./
+      ${PRJ_PATH}/noetl/api_config.py \
+      ${PRJ_PATH}/noetl/api_resolvers.py ./
 
 EXPOSE 8021
