+++ requirements.txt
@@ -1,7 +1,7 @@
 asyncio==3.4.3
 loguru==0.7.2
 aiofiles==23.2.1
-aiohttp==3.8.6
+aiohttp==3.9.1
 pyyaml==6.0
 requests>=2.31.0
 spacy==3.6.1
